# src/config/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TixUp API",
    "description": "API para gerenciamento de eventos e ingressos",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/api/eventos/": {
      "get": {
        "tags": [
          "Eventos"
        ],
        "summary": "Lista todos os eventos públicos",
        "description": "",
        "parameters": [
          {
            "name": "categoria",
            "in": "query",
            "description": "Filtrar por categoria",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "data",
            "in": "query",
            "description": "Filtrar por data (formato YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Eventos públicos listados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Evento"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Eventos"
        ],
        "summary": "Cria um novo evento",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Evento criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evento"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "example": "Show do Matuê"
                  },
                  "descricao": {
                    "type": "string",
                    "example": "Evento com atrações especiais"
                  },
                  "data": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-06-20"
                  },
                  "local": {
                    "type": "string",
                    "example": "Ginásio Central"
                  },
                  "preco": {
                    "type": "number",
                    "example": 99.9
                  },
                  "imagem": {
                    "type": "string",
                    "example": "https://imagem.com/img.jpg"
                  },
                  "publico": {
                    "type": "boolean",
                    "example": true
                  },
                  "categoria": {
                    "type": "string",
                    "example": "show"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/eventos/{id}/colaboradores": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "usuario_id": {
                    "example": "any"
                  },
                  "permissao": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/eventos/meus-eventos": {
      "get": {
        "tags": [
          "Eventos"
        ],
        "summary": "Lista os eventos do organizador logado",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/eventos/{id}": {
      "delete": {
        "tags": [
          "Eventos"
        ],
        "summary": "Deleta um evento por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do evento"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Evento deletado com sucesso"
          },
          "403": {
            "description": "Não autorizado a deletar este evento"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/eventos/{id}/vendas": {
      "get": {
        "tags": [
          "Vendas"
        ],
        "summary": "Lista todos os ingressos vendidos para um evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do evento"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de ingressos vendidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "uuid-do-ingresso"
                      },
                      "nome_comprador": {
                        "type": "string",
                        "example": "João Comprador"
                      },
                      "data_compra": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-18T10:00:00Z"
                      },
                      "status": {
                        "type": "string",
                        "example": "pendente"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Apenas organizadores podem visualizar vendas"
          },
          "404": {
            "description": "Evento não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/colaboradores/eventos/{id}/colaboradores": {
      "post": {
        "tags": [
          "Colaboradores"
        ],
        "summary": "Adiciona um colaborador a um evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do evento"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "usuario_id": {
                    "type": "string",
                    "example": "uuid-do-usuario"
                  },
                  "permissao": {
                    "type": "string",
                    "example": "checkin"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Colaboradores"
        ],
        "summary": "Lista os colaboradores de um evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do evento"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Evento": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "uuid"
                  }
                }
              },
              "nome": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "descricao": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "data": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date"
                  }
                }
              },
              "local": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "preco": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "imagem": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "publico": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "boolean"
                  }
                }
              },
              "categoria": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "organizador_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "uuid"
                  }
                }
              },
              "created_at": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              },
              "updated_at": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Token JWT no formato: Bearer <token>"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}